add: add circle, circlehook and useanimategrid
Circle.tsx extracts most of the logic from the useEffect hook of Box.tsx
and places it into a animateGrid function that is in the utils folder.
CircleHook takes all of the logic and state in Box and places it into
a useAnimateGrid hook. Added a types.ts file in models folder to store
reused types.