fix: move coords are relative to starting coords
The center coords of our element are now stored in state as screen
coordinates, as are the coords where we click the mouse. The vectorCalc
function now internally translates all coordinates such that the origin
is the center of the viewport. Using vector position, vector direction,
etc, means that we don't have to add a lot of logic to check the slope
of the line to decide whether the x and y ordinates need to be increased
or decreased.