fix: elements fully transition offscreen
Elements are translated offscreen based on the center of the element
and not the place where a user clicks. The value
t in f(t) = A + t*N is now set to the screen width, assuming the width
is wider than the height. All elements now travel a fixed distance and
so travel at the same velocity.