add: each element now contains part of an image
We import an image. In Content.tsx, we draw a 20x20 pixel image to a
20x20 pixel canvas element. We then write the data url for this to an
array. The elements of this array are then passed to our CircleHookClick
component and rendered as an image. As a result, we now have an image
that dynamically "explodes" when we click on the page!