refactor: move click functionality out of hooks
Added ContentRefactor, useAnimateGridClickRefactor and
TileHookClickRefactor components. Moved the event listeners from the
hook to ContentRefactor. ContentRefactor now has state for the click
coordinates. These are passed to TileHookClickRefactor and then passed
to the hook. Replaced getClickCoords with calcMoveCoords which takes
the click coordinates and calculates the coordinates to move each
element to.