add: toast message for first Loading
Cleaned up some code. Removed a ref from ContentRefactor. Moved the code
relating to the creation of our sub images into its own useEffect - it
was being rerun every time the user clicked and only needed to be run
once on load. Changed background to dark. Renamed all references to
cartesian or 'cart' to centerOrigin or center origin. Using cartesian
to refer to our use of a grid with the origin in the center of the
screen was a misunderstanding taken from a stack overflow question.